implement unreal engine 3 package reading support

+++ src/main/java/net/shrimpworks/unreal/packages/entities/Export.java
@@ -38,7 +38,7 @@ public abstract class Export implements Named {
 	 */
 	public final Name name;
 
-	public final int flags;
+	public final long flags;
 	public final int size;
 	public final int pos;
 
@@ -46,7 +46,7 @@ public abstract class Export implements Named {
 	private Set<Export> children;
 
 	Export(Package pkg, int index, ObjectReference classIndex, ObjectReference classSuperIndex, ObjectReference groupIndex, Name name,
-		   int flags, int size, int pos) {
+		   long flags, int size, int pos) {
 		this.pkg = pkg;
 		this.index = index;
 		this.classIndex = classIndex;
